+++ web_page_cryptids/data/explorer.py
@@ -1,5 +1,6 @@
-from .init import curs
+from .init import curs, IntegrityError
 from model.explorer import Explorer
+from ..errors import Missing, Duplicate
 
 curs.execute("""CREATE TABLE IF NOT EXISTS explorer(
     name TEXT PRIMARY KEY,
@@ -15,7 +16,11 @@ def model_to_dict(explorer: Explorer) -> dict:
 def get_one(name: str) -> Explorer:
     qry = "SELECT * FROM explorer WHERE name=:name;"
     params = {"name": name}
-    return row_to_model(curs.execute(qry, params).fetchone())
+    res = curs.execute(qry, params).fetchone()
+    if res:
+        return row_to_model(res)
+    else:
+        raise Missing(f"Explorer {name} not found")
 
 def get_all() -> list[Explorer]:
     qry = "SELECT * FROM explorer;"
@@ -24,20 +29,29 @@ def get_all() -> list[Explorer]:
 def create(explorer: Explorer) -> Explorer:
     qry = "INSERT INTO explorer VALUES (:name,:nationality);"
     params = model_to_dict(explorer)
-    curs.execute(qry, params)
+    try:
+        curs.execute(qry, params)
+    except IntegrityError:
+        raise Duplicate(f"Explorer {explorer.name} already exists")
     return get_one(explorer.name)
 
-def modify(explorer: Explorer) -> Explorer:
+def modify(name: str, explorer: Explorer) -> Explorer:
     qry = """UPDATE explorer
     SET name=:name,
     nationality=:nationality
     WHERE name=:name0;"""
     params = model_to_dict(explorer)
     params["name0"] = explorer.name
-    curs.execute(qry, params)
-    return get_one(explorer.name)
+    res = curs.execute(qry, params)
+    if res.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(f"Explorer {name} not found")
 
-def delete(explorer: Explorer) -> bool:
+def delete(name: str) -> bool:
+    if not name: return False
     qry = "DELETE FROM explorer WHERE name=:name;"
-    params = {"name": explorer.name}
-    return bool(curs.execute(qry, params))
+    params = {"name": name}
+    res = curs.execute(qry, params)
+    if res.rowcount != 1:
+        raise Missing(f"Explorer {name} not found")
